# Optimize `NaiveDate::add_days` for small values