I expect this method to be used much more often with small values than with values that end up an *x* number of years into the past or future. So I would like to add a fast path.

In the common case where the result of `add_checked_signed` and similar ends up within the same year we can avoid a couple of divisions and lookups and just change the ordinal.

Before:
```
bench_naivedate_add_signed
                        time:   [32.797 ns 34.745 ns 36.560 ns]
Found 13 outliers among 100 measurements (13.00%)
  13 (13.00%) high severe
```

After:
```
bench_naivedate_add_signed
                        time:   [12.886 ns 13.422 ns 13.855 ns]
                        change: [-65.054% -63.390% -61.655%] (p = 0.00 < 0.05)
                        Performance has improved.
```