@pitdicker 2023-08-28
> Does this have good test coverage?

> I'll have a look later today

That must have been one of the longer days in history...

With a small change in `add_days` to make the result invalid we have 37 failing tests:
```
failures:
    datetime::tests::test_add_sub_months
    datetime::tests::test_datetime_add_assign_local
    datetime::tests::test_datetime_add_days
    datetime::tests::test_datetime_add_months
    datetime::tests::test_datetime_date_and_time
    datetime::tests::test_datetime_fixed_offset
    datetime::tests::test_datetime_from_local
    datetime::tests::test_datetime_offset
    datetime::tests::test_datetime_rfc2822
    datetime::tests::test_datetime_rfc3339
    datetime::tests::test_datetime_sub_assign_local
    datetime::tests::test_datetime_sub_days
    datetime::tests::test_datetime_sub_months
    datetime::tests::test_from_system_time
    datetime::tests::test_parse_datetime_utc
    datetime::tests::test_rfc3339_opts
    datetime::tests::test_utc_datetime_from_str
    format::formatting::tests::test_datetime_format_alignment
    format::parse::tests::test_rfc2822
    format::parse::tests::test_rfc3339
    format::parsed::tests::issue_551
    format::parsed::tests::test_parsed_to_naive_date
    format::parsed::tests::test_parsed_to_naive_datetime_with_offset
    month::tests::test_month_enum_primitive_parse
    month::tests::test_month_enum_try_from
    naive::date::tests::test_date_add
    naive::date::tests::test_date_add_days
    naive::date::tests::test_date_parse_from_str
    naive::date::tests::test_date_sub_days
    naive::date::tests::test_naiveweek
    naive::date::tests::test_week_iterator_limit
    naive::date::tests::test_weeks_from
    naive::datetime::tests::test_and_local_timezone
    naive::datetime::tests::test_and_utc
    naive::datetime::tests::test_datetime_add
    offset::fixed::tests::test_date_extreme_offset
    offset::local::tests::verify_correct_offsets_distant_past
```

More seriously: we have reasonable but not exhaustive coverage for values within a year, for values that cross the year boundary and have to account for leap years, and for values than can cause an overflow in `naive::date::test_date_add_days`.

Adding more tests does not seem that useful to me.