@jtmoon79 2023-09-02
>> I recommend test cases for days values
>>    0 to 367 inclusive that compares add_days when `ADD_DAYS_ENABLE_FAST_PATH` is true and false

> Adding more tests does not seem that useful to me.

I think the confidence this provides is worth the relatively small amount of work. (admittedly, I'm not doing the work! though if you want me to then I can implement my suggestion)