@pitdicker 2023-09-11
@djc Did you want to give this a second look? Or was https://github.com/chronotope/chrono/pull/1214#discussion_r1285741322 only a question and is this approved?